Statistics ts 削除
